Add 404 page
This commit adds a 404 page when visiting a URL that does not resolve
to a VueJS component. The Page component has been updated so that
the title and subtitle are now static. We're only ever going to have
this title/version combo so we don't need it to be done in each view.

Signed-off-by: David Bond <davidsbond93@gmail.com>